add unit tests for flaskr app

+++ fancy_flaskr/flaskr/blueprints/blog.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, url_for, \
-        render_template, redirect, flash
+        render_template, redirect, flash, abort
 from flask_login import current_user, login_user, \
         logout_user, login_required
 
@@ -7,7 +7,7 @@ from ..models import Post
 from ..extension import db
 from ..forms import PostForm
 
-blog_bp = Blueprint('blog', __name__)
+blog_bp = Blueprint('blog', __name__, url_prefix='/post')
 
 
 @blog_bp.route('/')
@@ -36,6 +36,8 @@ def create():
 @login_required
 def update(id):
     post = Post.query.get_or_404(id)
+    if current_user.id != post.author_id:
+        abort(404)
     form = PostForm(title=post.title, body=post.body)
     if form.validate_on_submit():
         post.title = form.title.data
@@ -51,6 +53,8 @@ def update(id):
 @login_required
 def delete(id):
     post = Post.query.get_or_404(id)
+    if current_user.id != post.author_id:
+        abort(404)
     db.session.delete(post)
     db.session.commit()
     flash('Post deleted!')
